Refactor populate statistics method.
